Refactor authentication middleware and update user blocking/unblocking logic; enhance Next.js configuration and update dependencies
- Replaced `clerkMiddleware` with `authMiddleware` for improved route handling.
- Simplified blocking and unblocking user logic in `block.ts` and `follow.ts`.
- Updated `next.config.js` to use `remotePatterns` for image domains.
- Added new `next.config.ts` file for TypeScript support.
- Updated package dependencies in `package.json` and `package-lock.json`, including downgrading Clerk packages and upgrading Prisma.
- Refactored various components for better readability and performance, including layout and sidebar components.
- Added client directives to several components for improved rendering behavior.

+++ actions/block.ts
@@ -12,33 +12,22 @@ const roomService = new RoomServiceClient(
 );
 
 export const onBlock = async (id: string) => {
-  try {
-    const self = await getSelf();
+  const self = await getSelf();
 
-    let blockedUser;
+  if (self.id == id) return new Error("Cannot block yourself");
 
-    try {
-      await roomService.removeParticipant(self.id, id);
-      blockedUser = await blockUser(id);
-    } catch (error) {
-      throw new Error("internal error");
-    }
-    revalidatePath(`/u/${self.username}/community`);
+  await roomService.removeParticipant(self.id, id);
+  const blockedUser = await blockUser(id);
 
-    return blockedUser;
-  } catch (error) {
-    throw new Error("internal error");
-  }
+  revalidatePath(`/u/${self.username}/community`);
+
+  return blockedUser;
 };
 
 export const onUnblock = async (id: string) => {
-  try {
-    const self = await getSelf();
-    const unblockedUser = await unblockUser(id);
-    revalidatePath("/");
-    revalidatePath(`/u/${self.username}/community`);
-    return unblockedUser;
-  } catch (error) {
-    throw new Error("internal error");
-  }
+  const self = await getSelf();
+  const unblockedUser = await unblockUser(id);
+  revalidatePath("/");
+  revalidatePath(`/u/${self.username}/community`);
+  return unblockedUser;
 };
